Oilpan: Use HeapListHashSet for FontFaceList.
Getting rid of a few of our FIXME's that are easy to fix. Using
HeapListHashSet and not HeapLinkedHashSet because an iterator
is stored in a field and HeapLinkedHashSet is not safe for
modification during iteration. (HeapListHashSet cannot
survive removal of the element at the iterator, but is otherwise
OK with modification.)

R=erik.corry@gmail.com, sigbjornf@opera.com

Review URL: https://codereview.chromium.org/282263002

git-svn-id: svn://svn.chromium.org/blink/trunk@174274 bbb929c8-8fbe-4397-9dbb-9b2b20218538