Generate both toV8(RawPtr<X>) and toV8(X*) for [GarbageCollected] interfaces
This CL changes generated bindings so that both toV8(RawPtr<X>) and toV8(X*) are generated.
Otherwise, if someone uses RawPtr<X> instead of X* in the code base, it can be bound to
unexpected toV8() and a wrong wrapper can be returned to V8.
For more details, see the discussion in https://codereview.chromium.org/257503003/.

BUG=340522

Review URL: https://codereview.chromium.org/291293004

git-svn-id: svn://svn.chromium.org/blink/trunk@174473 bbb929c8-8fbe-4397-9dbb-9b2b20218538