Use a BitArray in RuntimeCSSEnabled instead of a Vector<bool>
Use a BitArray in RuntimeCSSEnabled instead of a Vector<bool> to keep track
of which features are enabled at runtime in order to use a little less memory.

This CL also adds a value argument to the WTF::BitArray constructor so that
bits can be initialized to 1 instead of 0,  as RuntimeCSSEnabled expects all
bits to be initially 1. A clear() method is also added so that bits can be
unset and not simply set.

R=eseidel@chromium.org, dstockwell@chromium.org

Review URL: https://codereview.chromium.org/290793004

git-svn-id: svn://svn.chromium.org/blink/trunk@174531 bbb929c8-8fbe-4397-9dbb-9b2b20218538