SVG: SVGAnimateElement should not cache |m_animatedElements|
Before this change, SVGAnimateElement cached resolved element instances in 
|m_animatedElements|. However, this is not updated synchronously when <use> 
shadow dom/instance tree is torn down or new instances are added.

The patch changes SVGAnimateElement to resolve its target instances on the fly 
when needed.

As |m_animatedElements| are no longer kept, we do not guarantee that 
animationStarted hook is called before each animation. animationEnded
hook is guaranteed to be called for elements stayed in the tree 
when the animation ended, but it may not be called for the animated element 
|removedFrom| the document tree while animation.

BUG=369860

Review URL: https://codereview.chromium.org/298873003

git-svn-id: svn://svn.chromium.org/blink/trunk@174923 bbb929c8-8fbe-4397-9dbb-9b2b20218538