Refine the UseCounters for DOMFocusIn/Out and focusin/out events
Hixie asked in bug 378163:
"Should we drop focusin/focusout?"
"Is there a compat need to support the focusin/focusout events?"

To answer that question, add a UseCounter for focusin/focusout.
Instead of counting all listeners, count cases where the event is
actually fired with a listener attached, which is a better estimate of
the compatibility risk. Make the same change to the DOMFocusInOut
counter to make the two comparable.

BUG=378163

Review URL: https://codereview.chromium.org/307613002

git-svn-id: svn://svn.chromium.org/blink/trunk@175014 bbb929c8-8fbe-4397-9dbb-9b2b20218538